feat: Allow to generate diff comment on GitHub PR.

+++ src/Uno.PackageDiff/Program.cs
@@ -1,48 +1,40 @@
 ﻿using Mono.Cecil;
-using Mono.Options;
-using NuGet.Common;
-using NuGet.Configuration;
 using NuGet.Packaging;
-using NuGet.Protocol;
-using NuGet.Protocol.Core.Types;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
-using System.Net;
-using System.Runtime.Serialization;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace Uno.PackageDiff
 {
-    class Program
+	class Program
 	{
-		private static readonly PackageSource NuGetOrgSource = new PackageSource("https://api.nuget.org/v3/index.json");
-
 		static async Task<int> Main(string[] args)
-        {
-			string sourceArgument = null;
-			string targetArgument = null;
-			string outputFile = null;
-			string diffIgnoreFile = null;
+		{
+			var context = new RunningContext(args);
+			var writer = BuildReportWriter(context);
+			var result = 1;
+			(string path, NuspecReader nuspecReader) source = default;
+			(string path, NuspecReader nuspecReader) target = default;
 
-			var p = new OptionSet() {
-				{ "base=", s => sourceArgument = s },
-				{ "other=", s => targetArgument = s },
-				{ "outfile=", s => outputFile = s },
-				{ "diffignore=", s => diffIgnoreFile = s },
-			};
-
-			p.Parse(args);
-
-			var source = await GetPackage(sourceArgument);
-			var target = await GetPackage(targetArgument);
-
+			if(writer is null)
+			{
+				return 1;
+			}
+			context.TryGetValue(CommandLineOptions.@base, out var sourceArgument);
+			context.TryGetValue(CommandLineOptions.diffignore, out var diffIgnoreFile);
 			try
 			{
+				source = await context.GetPackage(sourceArgument);
+				if(context.TryGetValue(CommandLineOptions.other, out var targetArgument))
+				{
+					target = await context.GetPackage(targetArgument);
+				}
+				else
+				{
+					target = await context.GetPackage(sourceArgument, true);
+				}
+
 				var sourcePlatforms = Directory.GetDirectories(Path.Combine(source.path, "lib"), "*.*", SearchOption.TopDirectoryOnly);
 				var targetPlatforms = Directory.GetDirectories(Path.Combine(target.path, "lib"), "*.*", SearchOption.TopDirectoryOnly);
 				var platformsFiles = from sourceAssemblyPath in sourcePlatforms
@@ -54,18 +46,20 @@ namespace Uno.PackageDiff
 										 Target = targetAssemblyPath
 									 };
 
+
 				var ignoreSet = DiffIgnore.ParseDiffIgnore(diffIgnoreFile, source.nuspecReader.GetVersion().ToString());
 
 				bool differences = false;
-				using (var writer = new StreamWriter(outputFile))
+				using(writer)
 				{
-					writer.WriteLine($"Comparison report for {source.nuspecReader.GetId()} **{source.nuspecReader.GetVersion()}** with {target.nuspecReader.GetId()} **{target.nuspecReader.GetVersion()}**");
+					writer.WriteHeader(source: (source.nuspecReader.GetId(), source.nuspecReader.GetVersion())
+						, target: (target.nuspecReader.GetId(), target.nuspecReader.GetVersion()));
 
 					foreach(var platform in platformsFiles)
 					{
-						writer.WriteLine($"# {platform.Platform} Platform");
+						writer.WritePlatform(platform.Platform);
 
-						foreach (var sourceFile in Directory.GetFiles(platform.Source, "*.dll"))
+						foreach(var sourceFile in Directory.GetFiles(platform.Source, "*.dll"))
 						{
 							var targetFile = Path.Combine(platform.Target, Path.GetFileName(sourceFile));
 
@@ -75,47 +69,60 @@ namespace Uno.PackageDiff
 
 								if(!ignoreSet.Assemblies.Any(m => m.FullName.Equals(targetFileName, StringComparison.OrdinalIgnoreCase)))
 								{
-									Console.WriteLine($"The assembly {targetFileName} could not be found in the target package");
+									context.Warn($"The assembly {targetFileName} could not be found in the target package");
 									differences = true;
 								}
 							}
 							else
 							{
-								Console.WriteLine($"Comparing {sourceFile} and {targetFile}");
+								context.Info($"Comparing {sourceFile} and {targetFile}");
 
-								differences |= CompareAssemblies(writer, sourceFile, targetFile, ignoreSet);
+								differences |= CompareAssemblies(context, writer, sourceFile, targetFile, ignoreSet);
 							}
 						}
 					}
 				}
 
 				var withDifferences = differences ? ", with differences" : "";
-				Console.WriteLine($"Done comparing{withDifferences}.");
+				context.Info($"Done comparing{withDifferences}.");
 
-				if (differences)
+				if(differences)
 				{
-					Console.WriteLine(@"Error : Build failed with unexpected differences. Modifications to the public API introduce binary breaking changes and should be avoided.");
-					Console.WriteLine("If these modifications were expected and intended, see https://github.com/nventive/Uno.PackageDiff#how-to-provide-an-ignore-set on how to ignore them.");
+					context.Error($"Build failed with unexpected differences. Modifications to the public API introduce binary breaking changes and should be avoided. If these modifications were expected and intended, see https://github.com/nventive/Uno.PackageDiff#how-to-provide-an-ignore-set on how to ignore them.");
 				}
 
-				return differences ? 1 : 0;
+				result = differences ? 1 : 0;
+			}
+			catch(Exception ex)
+			{
+				context.Error(ex.Message);
+				return 1;
 			}
 			finally
 			{
-				Directory.Delete(source.path, true);
-				Directory.Delete(target.path, true);
+				if(source != default)
+				{
+					Directory.Delete(source.path, true);
+				}
+
+				if(target != default)
+				{
+					Directory.Delete(target.path, true);
+				}
 			}
+
+			return result;
 		}
 
-		public static bool CompareAssemblies(StreamWriter writer, string sourceFile, string targetFile, IgnoreSet ignoreSet)
+		public static bool CompareAssemblies(RunningContext context, IReportWriter writer, string sourceFile, string targetFile, IgnoreSet ignoreSet)
 		{
 			using(var source = ReadModule(sourceFile))
 			using(var target = ReadModule(targetFile))
 			{
-				writer.WriteLine($"## {Path.GetFileName(sourceFile)}");
+				writer.WriteAssemblyName(Path.GetFileName(sourceFile));
 				var results = AssemblyComparer.CompareTypes(source, target);
 
-				return ReportAnalyzer.GenerateReport(writer, results, ignoreSet);
+				return ReportAnalyzer.GenerateReport(context, writer, results, ignoreSet);
 			}
 		}
 
@@ -126,79 +133,52 @@ namespace Uno.PackageDiff
 			return AssemblyDefinition.ReadAssembly(path, new ReaderParameters() { AssemblyResolver = resolver });
 		}
 
-		private static async Task<(string path, NuspecReader nuspecReader)> GetPackage(string packagePath)
+
+		private static IReportWriter BuildReportWriter(RunningContext context)
 		{
-			if(!packagePath.EndsWith(".nupkg", StringComparison.OrdinalIgnoreCase))
+			IReportWriter writer = null;
+			context.TryGetValue(CommandLineOptions.outtype, out var outputType);
+			if(string.IsNullOrEmpty(outputType))
 			{
-				var settings = Settings.LoadDefaultSettings(null);
-				var repositoryProvider = new SourceRepositoryProvider(settings, Repository.Provider.GetCoreV3());
-
-				var repository = repositoryProvider.CreateRepository(NuGetOrgSource, FeedType.HttpV3);
-
-				var searchResource = repository.GetResource<PackageSearchResource>();
-
-				var packages = (await searchResource.SearchAsync(packagePath, new SearchFilter(true, SearchFilterType.IsLatestVersion), skip: 0, take: 1000, log: new NullLogger(), cancellationToken: CancellationToken.None)).ToArray();
-
-				if(packages.Any())
+				writer = new Writers.MarkdownWriter(context);
+			}
+			else
+			{
+				var types = outputType.Split(",", StringSplitOptions.RemoveEmptyEntries);
+				if(types.Length == 0)
 				{
-					var latestStable = (await packages.First().GetVersionsAsync())
-						.OrderByDescending(v => v.Version)
-						.Where(v => !v.Version.IsPrerelease)
-						.FirstOrDefault();
-
-					if (latestStable != null)
-					{
-						var packageId = packagePath.ToLowerInvariant();
-						var version = latestStable.Version.ToNormalizedString().ToLowerInvariant();
-
-						// https://docs.microsoft.com/en-us/nuget/api/package-base-address-resource#download-package-content-nupkg
-						var url = $"https://api.nuget.org/v3-flatcontainer/{packageId}/{version}/{packageId}.{version}.nupkg";
-
-						Console.WriteLine($"Downloading {url}");
-						var outFile = Path.GetTempFileName();
-						await new WebClient().DownloadFileTaskAsync(new Uri(url), outFile);
-						return UnpackArchive(outFile);
-					}
-					else
-					{
-						throw new InvalidOperationException($"Unable to find stable {packagePath} in {NuGetOrgSource.SourceUri}");
-					}
+					writer = new Writers.MarkdownWriter(context);
 				}
 				else
 				{
-					throw new InvalidOperationException($"Unable to find {packagePath} in {NuGetOrgSource.SourceUri}");
+					var composite = new Writers.CompositeWriter();
+					foreach(var type in types)
+					{
+						if(string.Equals("md", type, StringComparison.OrdinalIgnoreCase))
+						{
+							composite.Add(new Writers.MarkdownWriter(context));
+						}
+						else if(string.Equals("diff", type, StringComparison.OrdinalIgnoreCase))
+						{
+							composite.Add(new Writers.XmlIgnoreWriter(context));
+						}
+						else if(string.Equals("github", type, StringComparison.OrdinalIgnoreCase))
+						{
+							composite.Add(new Writers.GitHubWriter(context));
+						}
+						else
+						{
+							context.Error($"Invalid outtype {type}.");
+						}
+					}
+					if(composite.Count == types.Length)
+					{
+						writer = composite;
+					}
 				}
 			}
-			else
-			{
-				return UnpackArchive(packagePath);
-			}
+			return writer;
 		}
 
-		private static (string path, NuspecReader nuspecReader) UnpackArchive(string packagePath)
-		{
-			var path = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString().Trim('{', '}'));
-
-			Directory.CreateDirectory(path);
-
-			Console.WriteLine($"Extracting {packagePath} -> {path}");
-			using (var file = File.OpenRead(packagePath))
-			{
-				using (var archive = new ZipArchive(file, ZipArchiveMode.Read))
-				{
-					archive.ExtractToDirectory(path);
-				}
-			}
-
-			if(Directory.GetFiles(path, "*.nuspec", SearchOption.AllDirectories).FirstOrDefault() is string nuspecFile)
-			{
-				return (path, new NuspecReader(nuspecFile));
-			}
-			else
-			{
-				throw new InvalidOperationException($"Unable to find nuspec file in {packagePath}");
-			}
-		}
-
-    }
+	}
 }
